init: setup storybook
