+++ apps/storybook/.storybook/preview.ts
@@ -1,14 +1,14 @@
-import type { Preview } from "@storybook/react";
+// import type { Preview } from "@storybook/react";
 
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
+// const preview: Preview = {
+//   parameters: {
+//     controls: {
+//       matchers: {
+//         color: /(background|color)$/i,
+//         date: /Date$/i,
+//       },
+//     },
+//   },
+// };
 
-export default preview;
+// export default preview;
