--- apps/storybook/src/pages/index.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <main>:O</main>;
-}
