+++ packages/ui/package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
+    "./my-button": "./src/my-button.tsx",
     ".": "./src/index.ts",
     "./*": [
       "./src/*.tsx",
@@ -12,6 +13,7 @@
     "./*/*": "./src/*/*.tsx"
   },
   "scripts": {
+    "build": "tsc",
     "lint": "eslint . --max-warnings 0",
     "generate:component": "turbo gen react-component"
   },
@@ -25,9 +27,12 @@
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
     "eslint": "^8.57.0",
+    "node-sass": "^9.0.0",
     "typescript": "^5.3.3"
   },
   "dependencies": {
+    "classnames": "^2.5.1",
+    "postcss": "^8",
     "react": "^18.2.0"
   },
   "typesVersions": {
