--- packages/ui/src/button.tsx
@@ -1,25 +1,0 @@
-"use client";
-
-import { ButtonHTMLAttributes, ReactNode } from "react";
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  children: ReactNode;
-  appName: string;
-}
-
-export const Button = ({
-  children,
-  className,
-  appName,
-  ...others
-}: ButtonProps) => {
-  return (
-    <button
-      className={className}
-      {...others}
-      onClick={() => alert(`Hello from your ${appName} app!`)}
-    >
-      {children}
-    </button>
-  );
-};
