+++ packages/ui/src/my-button.tsx
@@ -1,0 +1,8 @@
+export const MyButton = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div>
+      <h1>MyButton Component</h1>
+      {children}
+    </div>
+  );
+};
