+++ packages/ui/src/myButton/mybutton.module.scss
@@ -1,9 +1,13 @@
-@use "../utils/variables.module.scss";
+@use "../variables.module.scss";
 
 .button {
   background-color: variables.$primary-color;
   padding: 0.2rem;
+  &:hover {
+    cursor: pointer;
+  }
 }
+
 .clicked {
   background-color: variables.$accent-color;
 }
