+++ packages/ui/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "outDir": "dist"
   },
-  "include": ["src"],
+  "include": ["typings", "src"],
   "exclude": ["node_modules", "dist"]
 }
